requirements.txt
Faker==13.13.0
mypy==0.961
mypy-extensions==0.4.3
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
types-python-dateutil==2.8.17
typing_extensions==4.2.0
